refactor: renamed project files from Application Template to Met Art App
