+++ .idea/.name
@@ -1,1 +1,1 @@
-Application Template
+Met Art App
