+++ app/src/androidTest/java/com/example/metartapp/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.applicationtemplate
+package com.example.metartapp
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
@@ -19,6 +19,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.applicationtemplate", appContext.packageName)
+        assertEquals("com.example.metartapp", appContext.packageName)
     }
 }
