+++ app/src/main/java/com/example/metartapp/MainActivity.kt
@@ -1,9 +1,9 @@
-package com.example.applicationtemplate
+package com.example.metartapp
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
-import com.example.applicationtemplate.screens.home.HomeScreen
+import com.example.metartapp.screens.home.HomeScreen
 
 class MainActivity: ComponentActivity() {
