+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Application Template</string>
+    <string name="app_name">Met Art App</string>
 </resources>
